Saved searches emulate collection routes structure
