+++ app/main/posts/posts-module.js
@@ -93,13 +93,13 @@ angular.module('ushahidi.posts', [])
 .directive('collectionModeContext', require('./collections/mode-context.directive.js'))
 .directive('collectionEditor', require('./collections/editor.directive.js'))
 .directive('collectionListing', require('./collections/listing.directive.js'))
-
 .directive('collectionToggleButton', require('./common/collection-toggle/collection-toggle-button.js'))
 .directive('collectionToggleLink', require('./common/collection-toggle/collection-toggle-link.js'))
 
-
 .config(require('./posts-routes.js'))
 .config(require('./collections/collections-routes.js'))
+.config(require('./savedsearches/savedsearches-routes.js'))
+
 .run(['Leaflet', function (L) {
     L.Icon.Default.imagePath = '/img';
 }])
