+++ app/main/posts/savedsearches/savedsearches.html
@@ -1,8 +1,5 @@
 <div>
     <layout-class layout="{{layout}}"></layout-class>
     <saved-search-mode-context ng-if="layout === 'a'"></saved-search-mode-context>
-    <main role="main">
-        <div ui-view></div>
-        <!--<post-view-data ng-if="currentView() == 'data'" current-view="currentView" is-loading="isLoading" filters="filters"></post-view-data>-->
-    </main>
+    <div ui-view="savedsearchesData"></div>
 </div>
