+++ app/main/posts/views/main.html
@@ -29,6 +29,7 @@
         <!-- todo: toolbox-->
         <div ui-view noanimation="true"></div>
         <div ui-view="collectionsAll" noanimation="true"></div>
+        <div ui-view="savedsearchesAll" noanimation="true"></div>
         <ush-logo></ush-logo>
     </main>
 </div>
